# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "This endpoint allows a new user to register by providing their email, password, and other required information. Upon successful registration, the user will receive an access token and a refresh token.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation or business logic errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                },
                "examples": {
                  "EmailExists": {
                    "summary": "Email already exists",
                    "value": {
                      "errors": [
                        {
                          "message": "Email already exists"
                        }
                      ]
                    }
                  },
                  "PhoneExists": {
                    "summary": "Phone number already exists",
                    "value": {
                      "errors": [
                        {
                          "message": "Phone number already exists"
                        }
                      ]
                    }
                  },
                  "ValidationError": {
                    "summary": "Invalid input data",
                    "value": {
                      "errors": [
                        {
                          "message": "name should not be empty"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate user and receive access and refresh tokens",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid credentials or validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                },
                "examples": {
                  "InvalidCredentials": {
                    "summary": "Invalid credentials",
                    "value": {
                      "errors": [
                        {
                          "message": "Invalid credentials"
                        }
                      ]
                    }
                  },
                  "EmailRequired": {
                    "summary": "Email is required",
                    "value": {
                      "errors": [
                        {
                          "message": "email should not be empty"
                        }
                      ]
                    }
                  },
                  "PasswordRequired": {
                    "summary": "Password is required",
                    "value": {
                      "errors": [
                        {
                          "message": "password should not be empty"
                        }
                      ]
                    }
                  },
                  "InvalidEmail": {
                    "summary": "Invalid email format",
                    "value": {
                      "errors": [
                        {
                          "message": "email must be an email"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "description": "Generate new access and refresh tokens using a valid refresh token",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/refreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid token type or validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                },
                "examples": {
                  "InvalidTokenType": {
                    "summary": "Invalid token type",
                    "value": {
                      "errors": [
                        {
                          "message": "Invalid refresh token"
                        }
                      ]
                    }
                  },
                  "TokenRequired": {
                    "summary": "Refresh token is required",
                    "value": {
                      "errors": [
                        {
                          "message": "refreshToken should not be empty"
                        }
                      ]
                    }
                  },
                  "TokenMustBeString": {
                    "summary": "Refresh token must be a string",
                    "value": {
                      "errors": [
                        {
                          "message": "refreshToken must be a string"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                },
                "examples": {
                  "InvalidToken": {
                    "summary": "Invalid refresh token",
                    "value": {
                      "errors": [
                        {
                          "message": "Invalid refresh token"
                        }
                      ]
                    }
                  },
                  "ExpiredToken": {
                    "summary": "Expired refresh token",
                    "value": {
                      "errors": [
                        {
                          "message": "Invalid refresh token"
                        }
                      ]
                    }
                  },
                  "TokenNotFound": {
                    "summary": "Token not found in database",
                    "value": {
                      "errors": [
                        {
                          "message": "Invalid refresh token"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Internal server error"
                          }
                        },
                        "required": [
                          "message"
                        ]
                      }
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Airbnb Clone API",
    "description": "The Airbnb Clone API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "registerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "SecurePass123"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user",
            "example": "+1234567890"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "phone"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The unique identifier of the user",
            "example": "64b8f9e2c3a4f5d6e7f8g9h0"
          },
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the user",
            "example": "+1 (555) 123-4567"
          }
        },
        "required": [
          "_id",
          "name",
          "email",
          "phoneNumber"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "The user object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          },
          "accessToken": {
            "type": "string",
            "description": "The access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6ImpvaG4uZG9lQGV4YW1wbGUuY29tIiwiaWF0IjoxNjE3MzQ1NjAwLCJleHAiOjE2MTczNDkxMDAsInN1YiI6IjEifQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6ImpvaG4uZG9lQGV4YW1wbGUuY29tIiwiaWF0IjoxNjE3MzQ1NjAwLCJleHAiOjE2MTczNDkxMDAsInN1YiI6IjEifQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
      "loginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "password123"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "description": "actor role",
            "example": "user"
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6ImpvaG4uZG9lQGV4YW1wbGUuY29tIiwiaWF0IjoxNjE3MzQ1NjAwLCJleHAiOjE2MTczNDkxMDAsInN1YiI6IjEifQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJlbWFpbCI6ImpvaG4uZG9lQGV4YW1wbGUuY29tIiwiaWF0IjoxNjE3MzQ1NjAwLCJleHAiOjE2MTczNDkxMDAsInN1YiI6IjEifQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "refreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token used to generate new access and refresh tokens",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiaWF0IjoxNjg4NzQyMDAwLCJleHAiOjE2ODg3NDY4MDB9.abc123def456ghi789jkl012mno345pqr678stu901vwx234yz567890"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
    }
  }
}